Create packages/data-access; lift shared Repository interface (closes #292)
Creates @transformotion/data-access with Repository<T,ID>, RepositoryOptions,
QueryOptions, and PaginatedResult interfaces lifted from both apps'
base-repository.ts (byte-identical). Both apps' repositories/index.ts now
re-exports types from the package.
