+++ packages/data-access/src/index.ts
@@ -1,0 +1,1 @@
+export type { Repository, RepositoryOptions, QueryOptions, PaginatedResult } from './repository'
